pyproject.toml
[project]
name = "ollama-ipex"
version = "0.2.0"
description = "Local control plane for the IPEX-backed Ollama runtime."
readme = "README.md"
authors = [
    { name = "Ivann KAMDEM", email = "kapoivha@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.24.1",
]
license = "MIT"

[project.scripts]
ollama-ipex = "ipex.cli:main"
ipex = "ipex.cli:main"

[build-system]
requires = ["uv_build>=0.10.3,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "ipex"
